+++ perl/openprint/administrator_services.pm
@@ -26,8 +26,11 @@ sub edit {
 	} elsif ( $openprint::param{'btnFunction'} eq '>>' ) {
 		$Service = $Service->Next( {'category_id'=>$openprint::param{'ddmSearchCategory'}} );
 	} elsif ( $openprint::param{'btnFunction'} eq 'Delete' ) {
-		$Service->delete();
-		$Service = $Service->Next( {'category_id'=>$openprint::param{'ddmSearchCategory'}} );
+		foreach my $T ( openprint::Timetrack::find('service_id'=>$Service->id() ) ) {
+			$$variable{'error'} .= sprintf('Service is used in <a href="/timetrack/edit.html?timetrack_id=%1$d">Timetrack %1$d</a><br/>', $T->id() );
+		} # end foreach T
+		$$variable{'error'} .= $Service->delete() if ! $$variable{'error'};
+		$Service = $Service->Next( {'category_id'=>$openprint::param{'ddmSearchCategory'}} ) if ! $$variable{'error'};
 	} elsif ( $openprint::param{'btnFunction'} eq 'Save' ) {
 		if ( $openprint::param{'new_category'} ) {
 			if ( my @Categories = openprint::ServiceCategory::find('name'=>$openprint::param{'new_category'} ) ) {
